Fixing a misunderstanding I had about SSM.
The SSM spec states that a channel (Source,Group) is distinct from channel (Source',Group). I have not had the opportunity to test this as yet for Windows, Linux, or the other platforms that dotnet 5.0 runs on. That said, I also caution against relying on that aspect of the specification because all it takes is a single misimplementation and suddenly your code receives martian packets. It's best to code defensively, even for platforms that shouldn't need it.